requirements.txt
selenium==4.8.0
beautifulsoup4==4.11.2
openpyxl==3.1.0
python-dateutil==2.8.2
webdriver-manager==3.8.5
emoji==2.2.0
selenium_driverless
2captcha-python
django
djangorestframework